Previously removed code added back!
A bad comment lead to false identification of old code but it was an old
comment on new code! This has no been fixed by adding it back.

Also there was a small issue with the rsync where a recent mod for
CCP-EM added in an annoying blank entry to the rsync. This has now been
handled.