Issue #116
1. Fix for user reported annoyance of installing the hosts.conf in
secret, and also leaving the archive that is downloaded in $home. This
has now been replaced with a warning during install, and also the file
is created locally and not downloaded. 

2. Removed padding from version numbers so v1.01.005 > v1.1.5

3. Changed the README.md from markdown format to reStructured text, so
now the README is README.rst, this is to make Longbow more pip friendly
whilst still being bitbucket and possibly github (in case in future we
move).

4. setup.py modified to eliminate the python 3 issues, and also the long
description is now the README.rst, this means that users will see
something standard between the pip page and the bitbucket page, this
will also tie in with new pages on the HECBioSim website which will push
pip as the preffered way to install more prominently.

5. hosts.conf file has been removed from the development version since
this is now created on install for new users using pip, this will be
marked as deprecated on the website for a while so users on old versions
can still get it. Documentation will need to be updated to reflect this
change and provide the quickstart example that users can grab a copy of
if they are doing a more manual install. However pip will be the
encouraged way to install.