1. The way Longbow does logging is not friendly to library users, libraries should not be setting levels or formats on logs. Doing this forces library users to have to handle this.
Longbow was also logging to a single logger instance but formatted to
record the file that the error originated from. Since we are removing
the levels we needed to use a more natural way to perform logging (each
module gets a name that corresponds to its path within the package).

The logging for the application is now set within the Longbow
application main not in the main function (as this could be more useful
to soft library usage).

2. The console output messages have been aligned to 80 character widths
for portability

3. The need for the command-line parameter to be passed to main() has
been eliminated, this would have been an annoyance to library users.

4. The Longbow welcome message has been moved to the longbow main entry
point and out of the main() this means that library users will not got
the welcome message when linking to main() (they would normally supress
that)