The recovery function has been fleshed out with the code
The cwd parameter has been removed from the list of parameters

The NameError exception on the cleanup method within the longbow main
method has been moved to the inner loop within the cleanup method
itself. There would have been a case where if the name error was thrown
in a multijob scenario that files could be left on the remote host.