Modified the recovery file code to output the hosts file as a recovery file.
The recovery files are written to the .Longbow directory, I'm not sure
this is the best way to do this. Users of the application probably won't
notice but library users might suffer this way, so I have put an
override in the code that checks if the .Longbow dir exists first. If it
does then we will use it, if not then the recovery function is disabled.