Issue #32
Altered the code which performed a very weak version check based from
when only python 2.x was supported. This check now performs a check
based on a list of supported versions, this list should be updated with
newly supported versions each time unit tests are added for that version
and pass.