Issue #23
A file is now created inside the ~/.longbow directory in which the code
for the bash autocomplete is placed. An entry is then placed at the end
of the user bash profile to source this file. In future if it is
possible to provide uninstall scripts for pip (or if it is already) then
this can be removed upon uninstall.