Issue #25
Fix for bad user interupt behavior during a recovered session. A user
interrupt during a recovered session now results in all running/queued
jobs being aborted and all data downloaded as is. This is the same
behavior as a normal longbow session. The code however has created
duplication of the same code and tests, this should now be moved up a
level into the launcher method, this will result in some minor
refactoring so a new ticket should be filed.