Issue #24
Preparations for adding the cleanup code. This involved making the
recovery file name accessable outside of the scheduling module.

To accomplish this, the "recoveryfile" parameter was added to the
default data structure "jobs" this is initialised during the
configuration steps of Longbow. Unit tests were updated and some new
tests added to retain coverage and test new functionality.