Issue #4
Added new parameter "env-fix" to the jobs template and defaulted this to
false.

Added environment fix to sendtossh() which toggles on and off with
"env-fix" true/false

added logic to perform additional basic tests on the environment to
checkconnection() this will try and load the module system and if it is
missing then turn on environment fix.

Updated unit tests