Issue #63
Fix for keyerror when no modules are given with an absolute path to a
self compiled executable. Longbow will assume the user knows what they
are doing if they are able to compile software. If modules are required
such as intel etc then these should be given by the user in a job config
file.

This should have also fixed a related but slightly different and
unreported bug with the use of environment variables. Had a user given
an executable that didn't have a longbow plugin (ie a generic one) along
with no modules, this would also have given an error, even though they
might have configured the HPC user account with environment vars.