Issue #76
Adding in support for the slurm --gres flag. New configuration parameter
slurm-gres has been added and unit tests to verify this is working.

To make use of this parameter, the user should set the something like
the following in either the host (if all jobs require it) or job
configuration file

slurm-gres = gpu:1