Added module name override back
For codes where the module name on HPC is not the same as the package
name. For example if admins, have an unusual naming scheme etc. Then by
adding the code MODULENAME = "newname" to the package plugin will enable
custom names to be used. Note this is aimed at plugin developers and not
end users of Longbow. End users can simply provide the module name in a
configuration file.